+++ back/controllers/controllerRol.js
@@ -11,12 +11,16 @@ export const getAllRoles = async (req, res) => {
     }
 }
 
-export const getRole = async (req, res) => {
+export const getRole = async (req, res) => {    
     try {
-        const rol = await RolModel.findAll({
+        const [rol] = await RolModel.findAll({
             where: { idRol: req.params.idRol }
         });
-        res.json(rol[0]);
+        if(!rol)
+        {
+            return res.status(404).json({ message: "Rol no Registrado" });
+        }
+        res.json(rol);
     } catch (error) {
         res.json({ message: error.message });
     }
@@ -25,6 +29,13 @@ export const getRole = async (req, res) => {
 // Create role
 export const createRole = async (req, res) => {
     try {
+        const [rol] = await RolModel.findAll({
+            where: { Rol: req.body.Rol }
+        });
+        if(rol)
+        {
+            return res.json({ message: "Datos Duplicados: Ya Hay Un Registro Con Este Nombre" });
+        }
         const nuevoRol = await RolModel.create(req.body);
         const idRegistroMov = nuevoRol.idRol;
 
@@ -57,6 +68,13 @@ export const createRole = async (req, res) => {
 
 export const updateRole = async (req, res) => {
     try {
+        const [rol] = await RolModel.findAll({
+            where: { idRol: req.params.idRol }
+        });
+        if(!rol)
+        {
+            return res.status(404).json({ message: "Rol no Registrado" });
+        }
         await RolModel.update(req.body, {
             where: { idRol: req.params.idRol }
         });
@@ -77,6 +95,13 @@ export const updateRole = async (req, res) => {
 
 export const deleteRole = async (req, res) => {
     try {
+        const [rol] = await RolModel.findAll({
+            where: { idRol: req.params.idRol }
+        });
+        if(!rol)
+        {
+            return res.status(404).json({ message: "Rol no Registrado" });
+        }
         await RolModel.destroy({
             where: {
                 idRol: req.params.idRol
